Console has smart scoping
